refactor(class-service): moved events to different package, added abstraction, changed event payload types to String
