--- account/class-service/src/main/java/org/tutora/classservice/dto/ClassInvitationCreatedEvent.java
@@ -1,10 +1,0 @@
-package org.tutora.classservice.dto;
-
-import java.util.UUID;
-
-public record ClassInvitationCreatedEvent(
-        UUID senderId,
-        String className,
-        UUID receiverId
-) {
-}
