+++ account/class-service/src/main/java/org/tutora/classservice/event/ClassInvitationCreatedEvent.java
@@ -1,0 +1,8 @@
+package org.tutora.classservice.event;
+
+public record ClassInvitationCreatedEvent(
+        String senderId,
+        String className,
+        String receiverId
+) implements Event {
+}
