+++ account/class-service/src/main/java/org/tutora/classservice/event/Event.java
@@ -1,0 +1,4 @@
+package org.tutora.classservice.event;
+
+public interface Event {
+}
